chore: rename lib/common/stdpre.h => lib/common/StdPre.h
