+++ src/lib/common/StdDeque.h
@@ -6,7 +6,7 @@
  */
 #pragma once
 
-#include "stdpre.h"
+#include "StdPre.h"
 
 #include <deque>
