+++ src/lib/common/StdIStream.h
@@ -7,7 +7,7 @@
 
 #pragma once
 
-#include "stdpre.h"
+#include "StdPre.h"
 
 #if HAVE_ISTREAM
 #include <istream>
