+++ src/lib/common/StdOStream.h
@@ -7,7 +7,7 @@
 
 #pragma once
 
-#include "stdpre.h"
+#include "StdPre.h"
 
 #if HAVE_OSTREAM
 #include <ostream>
