+++ src/lib/common/stdset.h
@@ -7,7 +7,7 @@
 
 #pragma once
 
-#include "stdpre.h"
+#include "StdPre.h"
 
 #include <set>
